+++ lib/arachni/component/options/port.rb
@@ -3,24 +3,23 @@
     All rights reserved.
 =end
 
-###
-#
 # Network port option.
 #
-###
+# @author Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>
+# @abstract
 class Arachni::Component::Options::Port < Arachni::Component::Options::Base
+
+    def normalize
+        value.to_i
+    end
+
+    def valid?
+        return false if !super
+        (1..65535).include?( normalize )
+    end
+
     def type
         'port'
     end
 
-    def valid?( value )
-        return false if empty_required_value?( value )
-
-        if value && !value.to_s.empty? &&
-            ((!value.to_s.match( /^\d+$/ ) || value.to_i <= 0 || value.to_i > 65535))
-            return false
-        end
-
-        super
-    end
 end