+++ spec/support/fixtures/plugins/with_options.rb
@@ -11,9 +11,9 @@ class Arachni::Plugins::WithOptions < Arachni::Plugin::Base
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>',
             version:     '0.1',
             options:     [
-                Options::String.new( 'req_opt', [ true, 'Required option' ] ),
-                Options::String.new( 'opt_opt', [ false, 'Optional option' ] ),
-                Options::String.new( 'default_opt', [ false, 'Option with default value', 'value' ] )
+                Options::String.new( 'req_opt', required: true, description: 'Required option' ),
+                Options::String.new( 'opt_opt', description: 'Optional option' ),
+                Options::String.new( 'default_opt', description: 'Option with default value', default: 'value' )
             ]
         }
     end